requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=0.11,<2.0
Babel>=1.3

oslo.config>=1.11.0  # Apache-2.0
oslo.i18n>=1.5.0  # Apache-2.0
oslo.log>=1.2.0  # Apache-2.0
oslo.serialization>=1.4.0               # Apache-2.0
oslo.utils>=1.4.0                       # Apache-2.0

python-neutronclient